requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy
python-dotenv
anthropic      # only if you’ll use the LLM live
